+++ lib/src/ndialog_base.dart
@@ -23,9 +23,7 @@ class NDialog extends StatelessWidget {
   ///The (optional) set of actions that are displayed at the bottom of the dialog.
   final List<Widget>? actions;
 
-  const NDialog(
-      {Key? key, this.dialogStyle, this.title, this.content, this.actions})
-      : super(key: key);
+  const NDialog({Key? key, this.dialogStyle, this.title, this.content, this.actions}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -58,9 +56,7 @@ class NDialog extends StatelessWidget {
                               )
                       ],
                     ),
-                    style: (style.titleTextStyle ??
-                            (dialogTheme?.titleTextStyle)) ??
-                        (theme?.textTheme.titleLarge ?? TextStyle()),
+                    style: (style.titleTextStyle ?? (dialogTheme.titleTextStyle)) ?? (theme!.textTheme.titleLarge ?? TextStyle()),
                   ),
                 )
               : SizedBox.shrink(),
@@ -70,9 +66,7 @@ class NDialog extends StatelessWidget {
                     padding: style.contentPadding,
                     child: DefaultTextStyle(
                       child: Semantics(child: content),
-                      style: (style.contentTextStyle ??
-                              dialogTheme?.contentTextStyle) ??
-                          (theme?.textTheme.titleMedium ?? TextStyle()),
+                      style: (style.contentTextStyle ?? dialogTheme.contentTextStyle) ?? (theme?.textTheme.titleMedium ?? TextStyle()),
                     ),
                   ),
                 )
@@ -82,8 +76,7 @@ class NDialog extends StatelessWidget {
                   data: theme!.copyWith(
                     textButtonTheme: TextButtonThemeData(
                       style: TextButton.styleFrom(
-                        shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.zero),
+                        shape: RoundedRectangleBorder(borderRadius: BorderRadius.zero),
                       ),
                     ),
                   ),
@@ -96,9 +89,7 @@ class NDialog extends StatelessWidget {
                               children: List.generate(
                                 actions?.length ?? 0,
                                 (index) {
-                                  return Expanded(
-                                      child:
-                                          actions?[index] ?? SizedBox.shrink());
+                                  return Expanded(child: actions?[index] ?? SizedBox.shrink());
                                 },
                               ),
                             ),
@@ -125,8 +116,7 @@ class NDialog extends StatelessWidget {
     );
 
     return Padding(
-      padding: MediaQuery.of(context).viewInsets +
-          const EdgeInsets.symmetric(horizontal: 20.0, vertical: 24.0),
+      padding: MediaQuery.of(context).viewInsets + const EdgeInsets.symmetric(horizontal: 20.0, vertical: 24.0),
       child: Center(
         child: ConstrainedBox(
           constraints: const BoxConstraints(minWidth: 280.0),
@@ -136,12 +126,8 @@ class NDialog extends StatelessWidget {
             elevation: style.elevation,
             color: style.backgroundColor ?? theme?.dialogBackgroundColor,
             shape: style.borderRadius != null
-                ? RoundedRectangleBorder(
-                    borderRadius:
-                        style.borderRadius ?? BorderRadius.circular(5))
-                : style.shape ??
-                    RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(5.0)),
+                ? RoundedRectangleBorder(borderRadius: style.borderRadius ?? BorderRadius.circular(5))
+                : style.shape ?? RoundedRectangleBorder(borderRadius: BorderRadius.circular(5.0)),
           ),
         ),
       ),
@@ -243,15 +229,7 @@ class ZoomDialog extends DialogBackground {
   ///Action before dialog dismissed
   final Function? onDismiss;
 
-  const ZoomDialog(
-      {Key? key,
-      this.backgroundColor,
-      @required this.child,
-      this.initZoomScale = 0,
-      this.blur,
-      this.zoomScale = 3,
-      this.onDismiss})
-      : super(key: key);
+  const ZoomDialog({Key? key, this.backgroundColor, @required this.child, this.initZoomScale = 0, this.blur, this.zoomScale = 3, this.onDismiss}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -305,27 +283,10 @@ class DialogBackground extends StatelessWidget {
 
   final Color? barrierColor;
 
-  @Deprecated("Use barrierColor instead")
-  final Color? color;
-
-  const DialogBackground(
-      {Key? key,
-      this.dialog,
-      this.color,
-      this.dismissable,
-      this.blur,
-      this.onDismiss,
-      this.barrierColor})
-      : super(key: key);
+  const DialogBackground({Key? key, this.dialog, this.dismissable, this.blur, this.onDismiss, this.barrierColor}) : super(key: key);
 
   ///Show dialog directly
-  // Future show<T>(BuildContext context) => showDialog<T>(context: context, builder: (context) => this, barrierColor: barrierColor, barrierDismissible: dismissable ?? true);
-
-  Future<T?> show<T>(BuildContext context,
-          {DialogTransitionType? transitionType,
-          bool? dismissable,
-          Duration? transitionDuration}) =>
-      DialogUtils(
+  Future<T?> show<T>(BuildContext context, {DialogTransitionType? transitionType, bool? dismissable, Duration? transitionDuration}) => DialogUtils(
         child: this,
         dialogTransitionType: transitionType,
         dismissable: dismissable,
@@ -338,13 +299,13 @@ class DialogBackground extends StatelessWidget {
     return Material(
       type: MaterialType.canvas,
       color: Colors.transparent,
-      child: WillPopScope(
-        onWillPop: () async {
-          if (dismissable ?? true) {
-            if (onDismiss != null) onDismiss?.call();
-            return true;
+      child: PopScope(
+        canPop: dismissable ?? true,
+        onPopInvokedWithResult: (didPop, result) {
+          if (didPop) {
+            onDismiss?.call();
+            return;
           }
-          return false;
         },
         child: Stack(
           clipBehavior: Clip.antiAlias,
@@ -352,7 +313,7 @@ class DialogBackground extends StatelessWidget {
           fit: StackFit.expand,
           children: <Widget>[
             InkWell(
-              overlayColor: MaterialStateProperty.all(Colors.transparent),
+              overlayColor: WidgetStatePropertyAll(Colors.transparent),
               hoverColor: Colors.transparent,
               splashColor: Colors.transparent,
               highlightColor: Colors.transparent,
@@ -422,10 +383,8 @@ class DialogStyle {
     this.titleDivider = false,
     this.borderRadius = const BorderRadius.all(Radius.circular(10)),
     this.semanticsLabel = "",
-    this.titlePadding =
-        const EdgeInsets.only(left: 15.0, right: 15.0, top: 10.0),
-    this.contentPadding =
-        const EdgeInsets.only(right: 15.0, left: 15.0, top: 0.0, bottom: 15.0),
+    this.titlePadding = const EdgeInsets.only(left: 15.0, right: 15.0, top: 10.0),
+    this.contentPadding = const EdgeInsets.only(right: 15.0, left: 15.0, top: 0.0, bottom: 15.0),
     this.titleTextStyle,
     this.contentTextStyle,
     this.elevation = 24,